Move VecSkipError to a separate file, preparing for MapSkipError
